2 - Moving every file into a firefly package and separating the integrator into his own package
