+++ firefly/camera/__init__.py
@@ -1,0 +1,2 @@
+from firefly.type import Scene, Ray, SurfaceData
+from .orthographic_camera import *
